+++ hardwarelibrary/powerstrips/__init__.py
@@ -1,0 +1,6 @@
+from .capabilities import (
+    Capability,
+    OutletSwitchingControl, DefaultOutletControl, CurrentMeteringControl,
+)
+from .powerstripdevice import PowerStripDevice
+from .pwrusb import PwrUSBDevice, DebugPwrUSBDevice, DebugPwrUSBPort